Refactor StretchActionBase to its own source file
Move StretchActionBase implementation from BlockPolarStretchAction.cs to StretchActionBase.cs. Add [DxfSubClass] attribute, XML docs, and initialize all properties. Update BlockPolarStretchAction to override relevant properties and remove redundant base class code. Add necessary usings and namespace to new file.
